requirements.txt
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.6.0
astroid==2.14.2
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
black==23.1.0
bs4==0.0.1
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
comm==0.1.2
crispy-bootstrap4==2022.1
cron-descriptor==1.2.35
debugpy==1.6.6
decorator==5.1.1
dill==0.3.6
Django==4.1.6
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-crispy-forms==2.0
django-redis==5.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
executing==1.2.0
idna==3.4
ipykernel==6.21.1
ipython==8.9.0
isort==5.12.0
jedi==0.18.2
jupyter_client==8.0.2
jupyter_core==5.2.0
kombu==5.2.4
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathspec==0.11.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
py-edamam==0.2
Pygments==2.14.0
pylint==2.16.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.0.0
redis==4.5.1
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.3.0
wrapt==1.14.1
